refactor: remove duplicate GitExecutor implementation
- Remove old git/executor.rs that used tokio::process::Command directly
- Migrate all tests to use git_executor_adapter with CommandExecutor trait
- Fix test imports to use RealCommandExecutor instead of old GitExecutor
- Update module exports to remove old executor
- Consolidate to single testable git execution approach

This completes the migration to the new architecture where all git
operations go through the CommandExecutor abstraction for better
testability and consistency.
