+++ rust/src/git/mod.rs
@@ -1,7 +1,6 @@
 pub mod backend;
 pub mod command_backend;
 pub mod const_utils;
-pub mod executor;
 pub mod factory;
 pub mod git_executor_adapter;
 pub mod libs;
